Split contracts file to separate files for each class
